requirements.txt
torch
numpy
matplotlib
flask
flask-cors
pillow
pyyaml
tqdm
